+++ app/assets/javascripts/govuk_publishing_components/lib/initial-focus.js
@@ -1,0 +1,13 @@
+window.GOVUK = window.GOVUK || {}
+window.GOVUK.Modules = window.GOVUK.Modules || {};
+
+(function (Modules) {
+  function InitialFocus () { }
+
+  InitialFocus.prototype.start = function ($module) {
+    this.$module = $module[0]
+    this.$module.focus()
+  }
+
+  Modules.InitialFocus = InitialFocus
+})(window.GOVUK.Modules)